+++ vrp/src/main/java/info/uaic/vrp/Services/ProductService.java
@@ -51,7 +51,7 @@ public class ProductService {
                 product.setId(rs.getInt("id"));
                 product.setName(rs.getString("name"));
                 product.setDescription(rs.getString("description"));
-                product.setPrice(rs.getDouble("price"));
+                product.setPrice(rs.getBigDecimal("price"));
                 product.setWeight(rs.getDouble("weight"));
                 product.setStock(rs.getInt("quantity"));
                 products.add(product);
@@ -69,7 +69,7 @@ public class ProductService {
                     Product product = new Product();
                     product.setId(rs.getInt("id"));
                     product.setName(rs.getString("name"));
-                    product.setPrice(rs.getDouble("price"));
+                    product.setPrice(rs.getBigDecimal("price"));
                     return product;
                 }
             }
